Update Issuer Statistic structure. Refactor for a better organization.
Associated Issue: #3919
